fixing for betty
